+++ Edith_Lopez/NodeJs/ejercicio5.js
@@ -1,18 +1,13 @@
-var prompt= require ("prompt");
+var prompt = require ("prompt");
     prompt.start();
-    prompt.get(["NumberOne", "NumberTwo", "NumberThree"], function (err, result) {
-    console.log("Comand-line input received:");
-    var NumberOne = Number(result.NumberOne);
-    var NumberTwo = Number(result.NumberTwo);
-    var NumberThree = Number(result.NumberThree);
-
-    if(NumberOne===NumberTwo+ NumberOne) {
-      console.log("El numero uno es la suma de numero 2 y 3");
-    }else if (NumberTwo===NumberOne+NumberTrhree){
-      console.log("el numero dos es la suma de numero 1 y 3");
-    }else if (NumberTrhree===NumberOne+NumberTwo){
-      console.log("el numero tres es la suma de numero 1 y 2");
-    }else{
-      console.log("ninguno es la suma de los otros dos numeros");
-    }
+    prompt.get(["calificacion1", "calificacion2", "calificacion3", "Examen", "TrabajoFinal"], function (err, result) {
+    console.log("Comand-line imput received:");
+    var calificacion1 = Number(result.calificacion1);
+    var calificacion2 = Number(result.calificacion2);
+    var calificacion3 = Number(result.calificacion3);
+    var Examen = Number(result.Examen);
+    var TrabajoFinal = Number(result.TrabajoFinal);
+    var promedio = ((calificacion1+calificacion2+calificacion3)/3)*0.55;
+    var promedioFinal = (promedio + (Examen*0.30)+(TrabajoFinal*0.15))
+    console.log("Tu calificacion final es " +promedioFinal);
 });
